@@ -2,8 +2,11 @@ package at.qe.skeleton;
 
 import javax.faces.webapp.FacesServlet;
 
+import at.qe.skeleton.configs.CustomServletContextInitializer;
+import at.qe.skeleton.configs.WebSecurityConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
@@ -17,24 +20,20 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
  * to start this web application.
  *
  * This class is part of the skeleton project provided for students of the
- * courses "Software Architecture" and "Software Engineering" offered by the
- * University of Innsbruck.
+ * courses "Software Architecture" offered by the University of Innsbruck.
  */
 @SpringBootApplication
 @EnableGlobalMethodSecurity(prePostEnabled = true)
-/*
- * Prevent spring from trying to autowire the websocket-infrastructure: Exclude
- * the at.qe.skeleton.ui.websockets package from component scan.
- *
- * NOTE: Do not add any components to this package which should be managed by
- * spring. It is reserved for the CDI-injection-mechanisms (Weld). Only add
- * CDI-managed components.
- */
-@ComponentScan(basePackages = "at.qe.skeleton", excludeFilters = @Filter(type = FilterType.REGEX, pattern = "at.qe.skeleton.ui.websockets.*"))
+
 public class Main extends SpringBootServletInitializer {
 
     public static void main(String[] args) {
             SpringApplication.run(Main.class, args);
+    }
+    
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(new Class[]{Main.class, CustomServletContextInitializer.class, WebSecurityConfig.class});
     }
 
     @Bean
